+++ rubato/utils/rendering/surface.py
@@ -174,18 +174,6 @@ class Surface:
         else:
             raise ValueError(f"Position is outside of the ${self.__class__.__name__}.")
 
-    def get_pixel_tuple(self, pos: Vector | tuple[float, float]) -> tuple[int, int, int, int]:
-        """
-        Gets the color of a pixel on the surface.
-
-        Args:
-            pos: The position of the pixel.
-
-        Returns:
-            The color of the pixel.
-        """
-        return self.get_pixel(pos).to_tuple()
-
     def set_pixel(self, pos: Vector | tuple[float, float], color: Color = Color.black, blending: bool = True):
         """
         Draws a point on the surface.